requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
requests==2.27.1
beautifulsoup4==4.11.1
pandas==1.3.5
openpyxl==3.0.9
python-docx==0.8.11
numpy==1.21.5
nltk==3.91
scikit-learn==1.0.2
regex==2022.1.18
unicodedata2==14.0.0
lxml==4.8.0
spacy==3.8.0
joblib==1.1.0
# Mô hình tiếng Việt cho spaCy 3.8.x
https://github.com/trungtv/vi_spacy/raw/master/packages/vi_core_news_lg-3.8.0/dist/vi_core_news_lg-3.8.0-py3-none-any.whl 